- feat: NicknameComponent 파일로 별도 분리.
