+++ app/(pages)/mypage/page.tsx
@@ -2,6 +2,7 @@
 
 import { useState, useCallback } from 'react';
 import styled from 'styled-components';
+import NicnameComponent from '@/app/components/mypage/NicnameComponent';
 
 const ContainerMain = styled.div`
   display: flex;
@@ -50,21 +51,6 @@ const TitleStrong = styled.strong`
   font-size: 24px;
 `;
 
-// TODO: 디자인시 NicnameP, NicknameInput width 통일
-const NicknameP = styled.p`
-  display: inline-block;
-  width: 64px;
-`;
-
-const NicknameInput = styled.input`
-  display: inline-block;
-  width: 64px;
-`;
-
-const NicknameButton = styled.button`
-  display: inline-block;
-`;
-
 const TagP = styled.p`
   padding: 2px;
   text-align: center;
@@ -72,33 +58,6 @@ const TagP = styled.p`
   border-radius: 8px;
   background-color: #e0e0e0;
 `;
-
-const NicnameComponent = () => {
-  const [isModify, setIsModify] = useState(false);
-  const [value, setValue] = useState('닉네임');
-
-  const handleClickButton = useCallback(() => {
-    if (isModify) {
-      // TODO: 닉네임 수정 api 연동
-      console.log('닉네임 수정', value);
-    }
-
-    setIsModify((v) => !v);
-  }, [isModify, value]);
-
-  return (
-    <div>
-      {isModify ? (
-        <NicknameInput value={value} onChange={(e) => setValue(e.target.value)} />
-      ) : (
-        <NicknameP>{value}</NicknameP>
-      )}
-      <NicknameButton onClick={handleClickButton}>
-        {isModify ? '저장 아이콘' : '수정 아이콘'}
-      </NicknameButton>
-    </div>
-  );
-};
 
 export default function MyPage() {
   return (
